Shifted components of DataNode into new base class
